[CO-386] Separate and expose wallet actions
The action exposed there are necessary for someone which want to start a
wallet node from haskell. They're currently part of the server/Main.hs
which is an executable, making it impossible to be imported by another
module. This commits extract them from 'server/Main.hs' making them
easy to re-use and/or remove!
